requirements.txt
Flask==2.3.2
biopython==1.81